[AN-404] Fix tests (!53)
Reviewed-on: https://gitea.devos.club/antiq/typescript-llvm/pulls/53
Co-committed-by: eugene.makarov <eugene.makarov@myoffice.team>

+++ test/unit/cpp_integration/cpp/printable.cpp
@@ -2,5 +2,5 @@
 
 using namespace cpp;
 
-Printable::Printable(Point *point, string *s, int32_t i)
+Printable::Printable(Point *point, string *s, int64_t i)
     : point_(*point), string_(*s), i_(i) {}
